Extract shared logic to mixin
