+++ app/services/outgoings_types_updater.rb
@@ -1,5 +1,6 @@
 class OutgoingsTypesUpdater
   include Providers::Draftable
+  include TransactionTypesUpdateable
 
   attr_reader :params
 
@@ -27,24 +28,8 @@ class OutgoingsTypesUpdater
 
 private
 
-  def legal_aid_application
-    @legal_aid_application ||= LegalAidApplication.find_by(id: params[:legal_aid_application_id])
-  end
-
-  def legal_aid_application_params
-    params.require(:legal_aid_application).permit(transaction_type_ids: [])
-  end
-
   def transaction_types
     @transaction_types ||= TransactionType.debits.where(id: legal_aid_application_params[:transaction_type_ids])
-  end
-
-  def transaction_types_selected?
-    transaction_types.present?
-  end
-
-  def none_selected?
-    params[:legal_aid_application][:none_selected] == "true"
   end
 
   def synchronize_debit_transaction_types
@@ -56,10 +41,6 @@ private
     end
 
     destroy_all_debit_transaction_types(except: keep)
-  end
-
-  def add_transaction_type(transaction_type)
-    legal_aid_application.transaction_types << transaction_type
   end
 
   def destroy_all_debit_transaction_types(except:)
